Homemade-Pickle-main/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
boto3==1.34.0
botocore==1.34.0
